Toolbox Provider:
This provider exposes a resource which can be run during apply or destroy and reapplied based on changes.
We also allow for the outputs to be exposed back into terraform for reuse.
This provider is a last resort and you should favor a terraform provider which implements the needed functionality to reduce untracked dependencies.

The resource created is called 'external' and accessed as 'resource "toolbox_external" "name" {...}'

Changes based on:
  Date: 25-05-2023
  Terraform >= 1.3.6
  go >= 1.19.0
  Terraform Plugin Framework protocol = 6
  External Provider Commit Hash = 60cb348
  External Provider PR = 220

  Important as:
    go 1.19 fails by default when a program is executed in the work path.
    Terraform Plugin Framework changes returns null and "" as well as maps which are empty but were ommitted before. In the External providers data source, it is currently kept for backwards compatability but it is considered a regression and will eventually be removed.

Based on the External Provider by HashiCorp which used a data source to allow for external program access.
The issue with this is that it is required to run on every terraform run regardless of need and expected no side-effects.
We could use existing provisioners but we lose access to any outputs in our terraform state.